pyproject.toml
[project]
name = "epiplexity-curriculum-post-training"
version = "0.1.0"
description = "GRPO post-training with epiplexity-weighted curriculum sampling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2",
    "transformers>=4.40",
    "datasets>=2.18",
    "trl>=0.9",
    "accelerate>=0.29",
    "peft>=0.10",
    "sentencepiece>=0.1.99",
    "evaluate>=0.4",
    "PyYAML>=6.0",
]
